# Unroll the internal Webmachine monad.