--- a/stack-ghc7.8.yaml
+++ b/stack-ghc7.8.yaml
@@ -1,0 +1,8 @@
+flags: {}
+packages:
+- ./
+- example
+extra-deps:
+- microlens-0.3.4.1
+- http-types-0.9
+resolver: lts-2.2