This PR contains four optimizations to Airship's internal monad and tracing mechanisms, which (on my 2014 MBP) yield a **82%** gain in requests per second (from ~17000 to ~30000).

Before:
```
Running 30s test @ http://localhost:3000/
  1 threads and 8 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   480.45us  272.68us   9.77ms   92.86%
    Req/Sec    16.80k     1.66k   19.19k    73.67%
  501314 requests in 30.00s, 181.53MB read
Requests/sec:  16709.68
Transfer/sec:      6.05MB
```

After:
```
Running 30s test @ http://localhost:3000/
  1 threads and 8 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   250.78us   95.39us   5.48ms   77.80%
    Req/Sec    30.64k     1.02k   32.02k    83.00%
  914337 requests in 30.00s, 331.10MB read
Requests/sec:  30477.75
Transfer/sec:     11.04MB
```

The three optimizations are as follows:
1) We stop using `RWST`, as its `Writer` component leaks space. We replace it with an `RST` monad, extracted from the Snap framework, that provides assured strictness and an inlined `>>=` implementation. In addition, we use `modify'` rather than `modify` to ensure that the change is applied strictly (this yielded a small but significant increase in throughput).
2) We remove EitherT from our monad transformer, putting an Either inside the state tuple to represent failure.
3) We use `(:)` to add a new item onto the internal list of traces, rather than constructing a new list out of the item and using `mappend`. Though this constructs the list of traces in reverse order, we `reverse` said list before sending it to the client.
4) We use `ByteString` values in the internal trace list so as to avoid the overhead of an unnecessary `decodeUtf8` (since the header value is a `ByteString` anyway).

This significantly reduces the time we spend in `Airship.Internal.Decision.trace`. The code currently in master spends between 5-8% of its time in `trace`, as the following cost-center profile shows:

```
COST CENTRE              MODULE                        %time %alloc

utcTimeToRfc1123         Airship.Internal.Date           9.2    6.5
>>=                      Control.Monad.Trans.Either      7.0    5.0
trace                    Airship.Internal.Decision       5.5    9.9
lift                     Control.Monad.Trans.Either      5.3    7.5
o18                      Airship.Internal.Decision       4.6    6.2
fmap                     Airship.Types                   4.5    6.7
[ omitted ...]
```

After this optimization, it spends less time and memory in `trace`, `fmap`, and `lift` (inlining `trace` did not improve performance when compiled with -O2):
```
COST CENTRE              MODULE                        %time %alloc

COST CENTRE              MODULE                        %time %alloc
utcTimeToRfc1123         Airship.Internal.Date          13.4   15.5
trace                    Airship.Internal.Decision       5.3    6.0
socketConnection.sendall Network.Wai.Handler.Warp.Run    5.2    0.3
b13                      Airship.Internal.Decision       4.3    4.9
o18                      Airship.Internal.Decision       4.0    4.4
lift.\                   Airship.RST                     3.0    1.7
fmap.\                   Airship.RST                     2.9    4.9
state.\                  Airship.RST                     2.6    7.1
receiveloop              Network.Wai.Handler.Warp.Recv   2.2    0.0
traceHeader              Airship.Internal.Helpers        1.8    4.4
```

These tests were performed on the `basic` application provided in the Airship distribution, with the `wrk` tool: `wrk -t1 -c8 -d30s http://localhost:3000/`. The application itself was run with three cores, using `stack exec basic -- +RTS -N3 -RTS`. The memory footprint of the `basic` application remains constant at ~8MB.