@reiddraper 2016-07-28
I can confirm I am seeing a ~ 17% performance improvement compared to master (56cc3074c9ab3ff634db7806443cd983f024b2d9).